+++ src/main/kotlin/com/github/ivw/ezmode/EzModeAppService.kt
@@ -35,6 +35,8 @@ class EzModeAppService : Disposable {
     TypedAction.getInstance().ensureEzModeUnloaded()
 
     clearAllEditorsEzModeData()
+
+    restoreTabShortcuts()
   }
 
   companion object {