+++ src/main/kotlin/com/github/ivw/ezmode/config/TabKeyUtils.kt
@@ -3,6 +3,12 @@ package com.github.ivw.ezmode.config
 import com.intellij.openapi.actionSystem.*
 import com.intellij.openapi.keymap.*
 
+private val tabShortcut = KeyboardShortcut.fromString("TAB")
+private val newTabShortcut = KeyboardShortcut.fromString("alt T")
+
+val Keymap.ezModeUsesTab get() =
+  getShortcuts("ezmode.EnterEzMode").contains(tabShortcut)
+
 /**
  * Since tab is used to enter EzMode by default,
  * we move the normal tab/indent actions to another shortcut,
@@ -14,14 +20,20 @@ import com.intellij.openapi.keymap.*
  */
 fun moveTabShortcuts() {
   val keymap = KeymapManager.getInstance().activeKeymap
-  val tabShortcut = KeyboardShortcut.fromString("TAB")
-  val newTabShortcut = KeyboardShortcut.fromString("alt T")
-
-  val ezModeUsesTab = keymap.getShortcuts("ezmode.EnterEzMode").contains(tabShortcut)
-  if (ezModeUsesTab) {
+  if (keymap.ezModeUsesTab) {
     keymap.removeShortcut("EditorTab", tabShortcut)
     keymap.removeShortcut("EditorIndentSelection", tabShortcut)
     keymap.addShortcut("EditorTab", newTabShortcut)
     keymap.addShortcut("EditorIndentSelection", newTabShortcut)
+  }
+}
+
+fun restoreTabShortcuts() {
+  val keymap = KeymapManager.getInstance().activeKeymap
+  if (keymap.ezModeUsesTab) {
+    keymap.removeShortcut("EditorTab", newTabShortcut)
+    keymap.removeShortcut("EditorIndentSelection", newTabShortcut)
+    keymap.addShortcut("EditorTab", tabShortcut)
+    keymap.addShortcut("EditorIndentSelection", tabShortcut)
   }
 }